Moved remaining logic from RotatableTower to Tower
As all towers should now be able to rotate.

Also renamed `rotationDelay` to `rotationDuration`.
