--- Assets/Scripts/Towers/RotatableTowerLogic.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 40bdf2e92d29c3a4e8319c1f05e5334f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
