+++ Assets/Scripts/Towers/TowerLogic.cs
@@ -1,10 +1,17 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.VFX;
 
 
 public class TowerLogic : AbstractTower
 {
+    [SerializeField]
+    private GameObject directionalPointer = default;
+
+    [SerializeField]
+    private float rotationDuration = 200f;
+
     [ReadOnly, SerializeField]
     protected TurretInput turretInput;
 
@@ -12,6 +19,7 @@ public class TowerLogic : AbstractTower
     private TutorialText tutorialText = default;
 
     private RepairAnimationController repairAnimationController;
+    private int lastTweenId = -1;
 
     new void Awake()
     {
@@ -29,10 +37,32 @@ public class TowerLogic : AbstractTower
         tutorialText.SetButton(Direction.NORTH, true);
     }
 
+    void FixedUpdate()
+    {
+        ChangeDirection();
+    }
+
+    // Rotational movement using aim input
+    private void ChangeDirection()
+    {
+        if (!turretInput)
+            return;
+
+        Vector2 aim = turretInput.GetAimInput();
+        if (aim.sqrMagnitude > 0.01f)
+            RotateFacing(new Vector3(aim.x, 0f, aim.y), rotationDuration);
+    }
+
     // Allow turret to be operated when focused
     public override void Focus(PlayerStateController player)
     {
         turretInput = player.GetComponent<TurretInput>();
+
+        if (directionalPointer)
+        {
+            ScaleDirectionalPointer(Vector3.one);
+            directionalPointer.GetComponent<VisualEffect>().SetVector4("Color", player.FocusedColor);
+        }
 
         if (tutorialText)
             tutorialText.Focus();
@@ -43,8 +73,19 @@ public class TowerLogic : AbstractTower
     {
         turretInput = null;
 
+        if (directionalPointer)
+            ScaleDirectionalPointer(Vector3.zero);
+
         if (tutorialText)
             tutorialText.Unfocus();
+    }
+
+    private void ScaleDirectionalPointer(Vector3 toScale)
+    {
+        if (lastTweenId != -1)
+            LeanTween.cancel(lastTweenId);
+
+        lastTweenId = LeanTween.scale(directionalPointer, toScale, 0.15f).setEaseInOutQuad().id;
     }
 
     public override void Interact(PlayerStateController player)
